+++ R/listing.R
@@ -24,9 +24,19 @@ resolve_yaml_listing <- function(input_file, site_config, metadata, yaml_listing
 
   listing_articles <- list()
 
+  categories <- TRUE
+  categories_metadata <- TRUE
+
   for (collection in names(yaml_listing)) {
 
+
     collection <- site_collections(site_dir, site_config)[[collection]]
+
+    if (identical(collection[["categories"]], FALSE))
+      categories <- FALSE
+
+    if (identical(collection[["categories_metadata"]], FALSE))
+      categories_metadata <- FALSE
 
     articles <- yaml_listing[[collection$name]]
 
@@ -49,7 +59,10 @@ resolve_yaml_listing <- function(input_file, site_config, metadata, yaml_listing
   }
 
   # generate html
-  listing_html <- html_for_articles(listing_articles, caption = metadata$title, categories = TRUE)
+  listing_html <- html_for_articles(listing_articles,
+                                    caption = metadata$title,
+                                    categories = categories,
+                                    categories_metadata = categories_metadata)
 
   listing <- list(
     html = html_file(listing_html)
@@ -94,8 +107,9 @@ generate_listing <- function(input_file,
 
 article_listing_html <- function(site_dir, metadata, collection, articles) {
 
-  # detect whether we are showing categories in the sidebar
+  # detect whether we are showing categories (sidebar and inline)
   categories <- not_null(collection[["categories"]], TRUE)
+  categories_metadata <- not_null(collection[["categories_metadata"]], TRUE)
 
   # check for subscription
   subscription_html <- subscription_html(site_dir, collection)
@@ -117,11 +131,12 @@ article_listing_html <- function(site_dir, metadata, collection, articles) {
   html_for_articles(articles,
                     caption = metadata$title,
                     categories = categories,
+                    categories_metadata = categories_metadata,
                     subscription_html = subscription_html,
                     custom_html = custom_html)
 }
 
-html_for_articles <- function(articles, caption = NULL, categories = FALSE, subscription_html = NULL, custom_html = NULL) {
+html_for_articles <- function(articles, caption = NULL, categories = FALSE, categories_metadata = FALSE, subscription_html = NULL, custom_html = NULL) {
 
   # generate categories listing
   categories_html <- if (categories) categories_listing_html(articles)
@@ -138,11 +153,17 @@ html_for_articles <- function(articles, caption = NULL, categories = FALSE, subs
     else
       preview <- NULL
 
+    if (categories_metadata)
+      categories_metadata_html <- html_for_categories_metadata(article)
+    else
+      categories_metadata_html <- NULL
+
     a(href = article$path, class = "post-preview",
       tags$script(class = "post-metadata", type = "text/json",
                   HTML(jsonlite::toJSON(metadata))),
       div(class = "metadata",
-        div(class = "publishedDate", date_as_abbrev(article$date))
+        div(class = "publishedDate", date_as_abbrev(article$date)),
+        categories_metadata_html,
       ),
       div(class = "thumbnail", preview),
       div(class = "description",
@@ -195,6 +216,18 @@ html_for_articles <- function(articles, caption = NULL, categories = FALSE, subs
   }
 }
 
+html_for_categories_metadata <- function(article) {
+  if (is.list(article[["categories"]])) {
+    tags <- lapply(article[["categories"]], function(category) {
+      div(class = "tag", category)
+    })
+    div(class = "tags", tags)
+  } else {
+    NULL
+  }
+}
+
+
 custom_html <- function(site_dir, collection) {
 
   # check for custom HTML entry
@@ -246,7 +279,7 @@ categories_listing_html <- function(articles) {
       tags$ul(
         lapply(names(categories), function(name) {
           tags$li(
-            tags$a(href = paste0("#",gsub(" ", "_", name)), name),
+            tags$a(href = category_hash(name), name),
             tags$span(class = "category-count", sprintf("(%d)", categories[[name]]))
           )
         })
@@ -258,6 +291,10 @@ categories_listing_html <- function(articles) {
   }
 }
 
+
+category_hash <- function(category) {
+  paste0("#",gsub(" ", "_", category))
+}
 
 
 articles_info <- function(site_dir, collection) {